requirements.txt
asqlite==1.0.0
coloredlogs==15.0.1
discord.py==1.7.3
neotermcolor==2.0.10
psutil==5.8.0
pyfiglet==0.8.post1
python-dotenv==0.19.2
PyYAML==6.0
